+++ lib/src/auth/auth.dart
@@ -1,28 +1,33 @@
-part of mongo_dart;
+//part of mongo_dart;
+import 'package:mongo_dart/mongo_dart.dart' show Connection, Db, MongoDartError;
+import 'package:sasl_scram/sasl_scram.dart' show UsernamePasswordCredential;
 
-enum AuthenticationScheme { MONGODB_CR, SCRAM_SHA_1 }
+import 'mongodb_cr_authenticator.dart';
+import 'scram_sha1_authenticator.dart';
+import 'scram_sha256_authenticator.dart';
+
+enum AuthenticationScheme { MONGODB_CR, SCRAM_SHA_1, SCRAM_SHA_256 }
 
 abstract class Authenticator {
+  Authenticator();
+
+  factory Authenticator.create(AuthenticationScheme authenticationScheme, Db db,
+      UsernamePasswordCredential credentials) {
+    switch (authenticationScheme) {
+      case AuthenticationScheme.MONGODB_CR:
+        return MongoDbCRAuthenticator(db, credentials);
+      case AuthenticationScheme.SCRAM_SHA_1:
+        return ScramSha1Authenticator(credentials, db);
+      case AuthenticationScheme.SCRAM_SHA_256:
+        return ScramSha256Authenticator(credentials, db);
+      default:
+        throw MongoDartError("Authenticator wasn't specified");
+    }
+  }
+
   static String? name;
 
   Future authenticate(Connection connection);
-}
-
-Authenticator createAuthenticator(AuthenticationScheme authenticationScheme,
-    Db db, UsernamePasswordCredential credentials) {
-  switch (authenticationScheme) {
-    case AuthenticationScheme.MONGODB_CR:
-      return MongoDbCRAuthenticator(db, credentials);
-    case AuthenticationScheme.SCRAM_SHA_1:
-      return ScramSha1Authenticator(credentials, db);
-    default:
-      throw MongoDartError("Authenticator wasn't specified");
-  }
-}
-
-class UsernamePasswordCredential {
-  String? username;
-  String? password; // TODO: Encrypt this to secureString
 }
 
 abstract class RandomStringGenerator {
@@ -30,24 +35,6 @@ abstract class RandomStringGenerator {
       'ABCDEFGHIJKLMNOPQRSTUVWXYZ[\\]^_`abcdefghijklmnopqrstuvwxyz{|}~';
 
   String generate(int length);
-}
-
-class CryptoStrengthStringGenerator extends RandomStringGenerator {
-  @override
-  String generate(int length) {
-    var random = Random.secure();
-    var allowedCodeUnits = RandomStringGenerator.allowedCharacters.codeUnits;
-
-    var max = allowedCodeUnits.length - 1;
-
-    var randomString = <int>[];
-
-    for (var i = 0; i < length; ++i) {
-      randomString.add(allowedCodeUnits.elementAt(random.nextInt(max)));
-    }
-
-    return String.fromCharCodes(randomString);
-  }
 }
 
 Map<String, String> parsePayload(String payload) {