+++ lib/src/auth/scram_sha256_authenticator.dart
@@ -1,3 +1,4 @@
+//part of mongo_dart;
 import 'package:crypto/crypto.dart' as crypto;
 import 'package:sasl_scram/sasl_scram.dart'
     show
@@ -9,14 +10,14 @@ import 'package:mongo_dart/mongo_dart.dart' show Db;
 
 import 'sasl_authenticator.dart';
 
-class ScramSha1Authenticator extends SaslAuthenticator {
-  static String name = 'SCRAM-SHA-1';
+class ScramSha256Authenticator extends SaslAuthenticator {
+  static String name = 'SCRAM-SHA-256';
 
-  ScramSha1Authenticator(UsernamePasswordCredential credential, Db db)
+  ScramSha256Authenticator(UsernamePasswordCredential credential, Db db)
       : super(
             ScramMechanism(
-                'SCRAM-SHA-1', // Optionally choose hash method from a list provided by the server
-                crypto.sha1,
+                'SCRAM-SHA-256', // Optionally choose hash method from a list provided by the server
+                crypto.sha256,
                 credential,
                 CryptoStrengthStringGenerator()),
             db) {