+++ lib/src/database/commands/authentication_commands/sasl_continue_command/sasl_continue_options.dart
@@ -1,0 +1,12 @@
+
+/// SaslContinue command options;
+class SaslContinueOptions {
+  /// Undocumented command
+  /// Fields inferred from MongoDbDriver
+  /// [see also:](https://github.com/mongodb/specifications/blob/master/source/auth/auth.rst)
+
+  /// No Options at present
+  const SaslContinueOptions();
+
+  Map<String, Object> get options => <String, Object>{};
+}