+++ lib/src/network/connection_manager.dart
@@ -22,28 +22,61 @@ class _ConnectionManager {
 
   Future _connect(Connection connection) async {
     await connection.connect();
-    var isMasterCommand = DbCommand.createIsMasterCommand(db);
-    var replyMessage = await connection.query(isMasterCommand);
-    if (replyMessage.documents == null || replyMessage.documents!.isEmpty) {
-      throw MongoDartError('Empty reply message received');
+    var result = <String, Object?>{keyOk: 0.0};
+    // As I couldn't set-up a pre 3.6 environment, I check not only for
+    // a {ok: 0.0} but also for any other error
+    try {
+      var helloCommand = HelloCommand(db,
+          username: connection.serverConfig.userName, connection: connection);
+      result = await helloCommand.execute(skipStateCheck: true);
+    } catch (e) {
+      //Do nothing
     }
-    var documents = replyMessage.documents!;
-    _log.fine(() => documents.first.toString());
-    var master = documents.first['ismaster'] == true;
-    connection.isMaster = master;
-    if (master) {
-      _masterConnection = connection;
-      MongoModernMessage.maxBsonObjectSize =
-          documents.first[keyMaxBsonObjectSize];
-      MongoModernMessage.maxMessageSizeBytes =
-          documents.first[keyMaxMessageSizeBytes];
-      MongoModernMessage.maxWriteBatchSize =
-          documents.first[keyMaxWriteBatchSize];
+    if (result[keyOk] == 1.0) {
+      var resultDoc = HelloResult(result);
+      var master = resultDoc.isWritablePrimary;
+      connection.isMaster = master;
+      if (master) {
+        _masterConnection = connection;
+        MongoModernMessage.maxBsonObjectSize = resultDoc.maxBsonObjectSize;
+        MongoModernMessage.maxMessageSizeBytes = resultDoc.maxMessageSizeBytes;
+        MongoModernMessage.maxWriteBatchSize = resultDoc.maxWriteBatchSize;
+      }
+      connection.serverCapabilities.getParamsFromHello(resultDoc);
+      if (db._authenticationScheme == null &&
+          resultDoc.saslSupportedMechs != null) {
+        if (resultDoc.saslSupportedMechs!.contains('SCRAM-SHA-256')) {
+          db._authenticationScheme = AuthenticationScheme.SCRAM_SHA_256;
+        } else if (resultDoc.saslSupportedMechs!.contains('SCRAM-SHA-1')) {
+          db._authenticationScheme = AuthenticationScheme.SCRAM_SHA_1;
+        }
+      }
+    } else {
+      var isMasterCommand = DbCommand.createIsMasterCommand(db);
+      var replyMessage = await connection.query(isMasterCommand);
+      if (replyMessage.documents == null || replyMessage.documents!.isEmpty) {
+        throw MongoDartError('Empty reply message received');
+      }
+      var documents = replyMessage.documents!;
+      _log.fine(() => documents.first.toString());
+      var master = documents.first['ismaster'] == true;
+      connection.isMaster = master;
+      if (master) {
+        _masterConnection = connection;
+        MongoModernMessage.maxBsonObjectSize =
+            documents.first[keyMaxBsonObjectSize];
+        MongoModernMessage.maxMessageSizeBytes =
+            documents.first[keyMaxMessageSizeBytes];
+        MongoModernMessage.maxWriteBatchSize =
+            documents.first[keyMaxWriteBatchSize];
+      }
+      connection.serverCapabilities.getParamsFromIstMaster(documents.first);
     }
-    connection.serverCapabilities.getParamsFromIstMaster(documents.first);
 
     if (db._authenticationScheme == null) {
-      if (connection.serverCapabilities.maxWireVersion >= 3) {
+      if ((connection.serverCapabilities.fcv?.compareTo('4.0') ?? -1) > -1) {
+        db._authenticationScheme = AuthenticationScheme.SCRAM_SHA_256;
+      } else if (connection.serverCapabilities.maxWireVersion >= 3) {
         db._authenticationScheme = AuthenticationScheme.SCRAM_SHA_1;
       } else {
         db._authenticationScheme = AuthenticationScheme.MONGODB_CR;
@@ -58,6 +91,25 @@ class _ConnectionManager {
             connection: connection);
         _log.fine(() => '$db: ${connection.serverConfig.hostUrl} connected');
       } catch (e) {
+        /// Atlas does not currently support SHA_256
+        if (e is MongoDartError &&
+            e.mongoCode == 8000 &&
+            e.errorCodeName == 'AtlasError' &&
+            e.message.contains('SCRAM-SHA-256') &&
+            db._authenticationScheme == AuthenticationScheme.SCRAM_SHA_256) {
+          _log.warning(() => 'Atlas connection: SCRAM_SHA_256 not available, '
+              'downgrading to SCRAM_SHA_1');
+          db._authenticationScheme = AuthenticationScheme.SCRAM_SHA_1;
+          try {
+            await db.authenticate(connection.serverConfig.userName!,
+                connection.serverConfig.password ?? '',
+                connection: connection);
+            _log.fine(
+                () => '$db: ${connection.serverConfig.hostUrl} connected');
+          } catch (e) {
+            rethrow;
+          }
+        }
         if (connection == _masterConnection) {
           _masterConnection = null;
         }